+++ lib/RpcConnector.php
@@ -28,6 +28,10 @@ use SimpleSAML\Logger;
  */
 class RpcConnector
 {
+    const COOKIE_FILE = '/tmp/proxyidp_cookie.txt';
+    const CONNECT_TIMEOUT = 1;
+    const TIMEOUT = 15;
+
     private $rpcUrl;
     private $user;
     private $password;
@@ -57,6 +61,10 @@ class RpcConnector
         curl_setopt($ch, CURLOPT_URL, $uri . '?' . $paramsQuery);
         curl_setopt($ch, CURLOPT_USERPWD, $this->user . ":" . $this->password);
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+        curl_setopt($ch, CURLOPT_COOKIEJAR, self::COOKIE_FILE);
+        curl_setopt($ch, CURLOPT_COOKIEFILE, self::COOKIE_FILE);
+        curl_setopt($ch, CURLOPT_CONNECTTIMEOUT, self::CONNECT_TIMEOUT);
+        curl_setopt($ch, CURLOPT_TIMEOUT, self::TIMEOUT);
 
         $json = curl_exec($ch);
         curl_close($ch);
@@ -94,6 +102,10 @@ class RpcConnector
             ['Content-Type:application/json', 'Content-Length: ' . strlen($paramsJson)]
         );
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+        curl_setopt($ch, CURLOPT_COOKIEJAR, self::COOKIE_FILE);
+        curl_setopt($ch, CURLOPT_COOKIEFILE, self::COOKIE_FILE);
+        curl_setopt($ch, CURLOPT_CONNECTTIMEOUT, self::CONNECT_TIMEOUT);
+        curl_setopt($ch, CURLOPT_TIMEOUT, self::TIMEOUT);
 
         $json = curl_exec($ch);
         curl_close($ch);